+++ crates/teloxide/src/dispatching/dispatcher.rs
@@ -10,7 +10,12 @@ use crate::{
 };
 
 use dptree::di::{DependencyMap, DependencySupplier};
-use futures::{future::BoxFuture, stream::FuturesUnordered, StreamExt};
+use either::Either;
+use futures::{
+    future::{self, BoxFuture},
+    stream::FuturesUnordered,
+    FutureExt as _, StreamExt as _,
+};
 use tokio_stream::wrappers::ReceiverStream;
 
 use std::{
@@ -19,6 +24,7 @@ use std::{
     future::Future,
     hash::Hash,
     ops::{ControlFlow, Deref},
+    pin::pin,
     sync::{
         atomic::{AtomicBool, AtomicU32, Ordering},
         Arc,
@@ -271,7 +277,6 @@ where
     ///  - An update from Telegram;
     ///  - [`crate::types::Me`] (can be used in [`HandlerExt::filter_command`]).
     ///
-    /// [`shutdown`]: ShutdownToken::shutdown
     /// [`HandlerExt::filter_command`]: crate::dispatching::HandlerExt::filter_command
     pub async fn dispatch(&mut self)
     where
@@ -289,19 +294,39 @@ where
     /// `update_listener_error_handler`.
     ///
     /// This method adds the same dependencies as [`Dispatcher::dispatch`].
-    ///
-    /// [`shutdown`]: ShutdownToken::shutdown
     pub async fn dispatch_with_listener<'a, UListener, Eh>(
         &'a mut self,
-        mut update_listener: UListener,
+        update_listener: UListener,
         update_listener_error_handler: Arc<Eh>,
     ) where
         UListener: UpdateListener + 'a,
         Eh: ErrorHandler<UListener::Err> + 'a,
         UListener::Err: Debug,
     {
+        self.try_dispatch_with_listener(update_listener, update_listener_error_handler)
+            .await
+            .expect("Couldn't prepare dispatching context")
+    }
+
+    /// Same as `dispatch_with_listener` but returns a `Err(_)` instead of
+    /// panicking when the initial telegram api call (`get_me`) fails.
+    ///
+    /// Starts your bot with custom `update_listener` and
+    /// `update_listener_error_handler`.
+    ///
+    /// This method adds the same dependencies as [`Dispatcher::dispatch`].
+    pub async fn try_dispatch_with_listener<'a, UListener, Eh>(
+        &'a mut self,
+        mut update_listener: UListener,
+        update_listener_error_handler: Arc<Eh>,
+    ) -> Result<(), R::Err>
+    where
+        UListener: UpdateListener + 'a,
+        Eh: ErrorHandler<UListener::Err> + 'a,
+        UListener::Err: Debug,
+    {
         // FIXME: there should be a way to check if dependency is already inserted
-        let me = self.bot.get_me().send().await.expect("Failed to retrieve 'me'");
+        let me = self.bot.get_me().send().await?;
         self.dependencies.insert(me);
         self.dependencies.insert(self.bot.clone());
 
@@ -321,15 +346,22 @@ where
             loop {
                 self.remove_inactive_workers_if_needed().await;
 
-                tokio::select! {
-                    upd = stream.next() => match upd {
-                        None => break,
+                let res = future::select(stream.next(), pin!(self.state.wait_for_changes()))
+                    .map(either)
+                    .await
+                    .map_either(|l| l.0, |r| r.0);
+
+                match res {
+                    Either::Left(upd) => match upd {
                         Some(upd) => self.process_update(upd, &update_listener_error_handler).await,
+                        None => break,
                     },
-                    () = self.state.wait_for_changes() => if self.state.is_shutting_down() {
-                        if let Some(token) = stop_token.take() {
-                            log::debug!("Start shutting down dispatching...");
-                            token.stop();
+                    Either::Right(()) => {
+                        if self.state.is_shutting_down() {
+                            if let Some(token) = stop_token.take() {
+                                log::debug!("Start shutting down dispatching...");
+                                token.stop();
+                            }
                         }
                     }
                 }
@@ -347,6 +379,7 @@ where
             .await;
 
         self.state.done();
+        Ok(())
     }
 
     async fn process_update<LErr, LErrHandler>(
@@ -449,9 +482,8 @@ where
         }
     }
 
-    /// Setups the `^C` handler that [`shutdown`]s dispatching.
-    ///
-    /// [`shutdown`]: ShutdownToken::shutdown
+    /// Setups the `^C` handler in order to call [`ShutdownToken::shutdown`]
+    /// when pressed.
     #[cfg(feature = "ctrlc_handler")]
     #[deprecated(since = "0.10.0", note = "use `enable_ctrlc_handler` on builder instead")]
     pub fn setup_ctrlc_handler(&mut self) -> &mut Self {
@@ -459,8 +491,8 @@ where
         self
     }
 
-    /// Returns a shutdown token, which can later be used to shutdown
-    /// dispatching.
+    /// Returns a shutdown token, which can later be used to
+    /// [`ShutdownToken::shutdown`].
     pub fn shutdown_token(&self) -> ShutdownToken {
         self.state.clone()
     }
@@ -578,6 +610,12 @@ async fn handle_update<Err>(
     }
 }
 
+fn either<L, R>(x: future::Either<L, R>) -> Either<L, R> {
+    match x {
+        future::Either::Left(l) => Either::Left(l),
+        future::Either::Right(r) => Either::Right(r),
+    }
+}
 #[cfg(test)]
 mod tests {
     use std::convert::Infallible;